--- backtrace-library/src/main/java/backtraceio/library/common/SharePreferencesManager.java
@@ -1,4 +1,0 @@
-package backtraceio.library.common;
-
-public class SharePreferencesManager {
-}
